+++ crates/debugger_ui/src/lib.rs
@@ -1,6 +1,6 @@
 use dap::debugger_settings::DebuggerSettings;
+use debug_session::DebugSession;
 use debugger_panel::{DebugPanel, ToggleFocus};
-use debugger_panel_item::DebugSession;
 use gpui::App;
 use settings::Settings;
 use workspace::{
@@ -10,8 +10,8 @@ use workspace::{
 
 pub mod attach_modal;
 pub mod console;
+pub mod debug_session;
 pub mod debugger_panel;
-pub mod debugger_panel_item;
 pub mod loaded_source_list;
 pub mod module_list;
 pub mod stack_frame_list;
